requirements.txt
pyodbc==4.0.34
python-decouple==3.6